Rename back to sideshow.js
