+++ bin/sideshow.js
